SE再生の機能をModelクラス側に持っていく/その他リファクタ
